Use input capture in main sketch
The main sketch was slightly modified so that it can use the new input
library, which tries to defragment the ways that input is sent to the
Arduino. Previously, there was a seperate library for getting data from an
XBee. Now, all the ways to give input to the Arduino is within a single
library. A new feature includes using backspace function to make
corrections. This simplifies the computation that the Arduino must do,
especially when taking input from the keypad.