Add better selectivity to respond to keypresses
The function to capture a keypress was modified to be more selective when
determining what type of key was pressed and what action should follow.
This was mostly used to better differentiate between a RESET trigger or a
CANCEL trigger.

The main sketch was modified to better handle the displaying of status
messages.