Capture dSPACE serial data through XBee
The input capture library files were modified to include the methods for
communicating with dSPACE through an XBee module.

The main sketch was modified so that the method for receiving serial
inputs through the XBee was disabled. It was causing conflicts with the
dSPACE methods. Further refinement is needed to resolve this conflict.