Add balanced input and profile modes
Added a dSPACE balanced inupt mode. This mode allows a user to change the
load to a balanced state with a single input, corresponding to the number
of relays that should be ON with an offset.

Added another dSPACE mode, profile mode, that allows the user to run a
predefined load profile. Future implementation will allow the user to
input the profile parameters.