Add style for clang-format
Introduced the 'clang-format' tool to ease code formatting and handling
of white-space. The added file contains a formatting style based on the
pre-defined C++ Google style guide. The only difference is that #include
directives are not sorted. This is due to possible conflicts with the
F() macro used in unit test files and in the Google Test library.