Refactor test for collect simple dSPACE input
Also includes a function that sets the expectations for the handling of
a dSPACE transmission for that input. This function reduces the
verbosity that comes with handling a transmission. In future upgrades,
this handling and implementation may change in order to simplify this
process.

Some of the older tests were removed because the functionality they were
testing is covered in the new test.