Rename relay state function
A global relay state is needed to keep track of previous relay values.
Without this, a shortened input sequence cannot produce a complete state
for the relays. For example, if the input were "AB4" then we know that
phases A and B will have 4 relays ON in the new state. But there is no
indication of the previous values for phase C and DC relays, thus we
cannot form a complete new relay state. This requires a rename and
different signature for the relay state generator function.